make fuel-indexer-lib standalone
