+++ packages/fuel-indexer-macros/src/helpers.rs
@@ -5,8 +5,7 @@ use async_graphql_parser::types::{BaseType, FieldDefinition, Type, UnionType};
 use async_graphql_value::Name;
 use fuel_abi_types::abi::program::{ProgramABI, TypeDeclaration};
 use fuel_indexer_database_types::*;
-use fuel_indexer_lib::{graphql::*, ExecutionSource};
-use fuel_indexer_types::type_id;
+use fuel_indexer_lib::{graphql::*, type_id, ExecutionSource};
 use fuels_code_gen::utils::Source;
 use proc_macro2::TokenStream;
 use quote::{format_ident, quote};
