+++ packages/fuel-indexer-types/src/receipt.rs
@@ -1,7 +1,8 @@
 use crate::{
     scalar::{Address, AssetId, Bytes32, ContractId, MessageId, Nonce},
-    type_id, TypeId, FUEL_TYPES_NAMESPACE,
+    TypeId, FUEL_TYPES_NAMESPACE,
 };
+use fuel_indexer_lib::type_id;
 use serde::{Deserialize, Serialize};
 
 #[derive(Deserialize, Serialize, Debug, Clone)]
